* 缓存过期-跳表实现策略  *  * 优点：定时删除时不用做过多消耗，支持惰性删除  * 实现方法：将expireMap的value作为跳表节点，跳表支持自定义层数用于平衡空间和时间，通过key可以直接拿到跳表节点，  * 删除可直接更改指针到待删除链表中惰性删除

+++ cache-core/src/main/java/com/github/houbb/cache/core/support/util/SkipNode.java
@@ -24,7 +24,7 @@ public interface SkipNode<V> {
      * 获取对应的 key 信息
      * @since 0.0.13
      */
-    int getKey();
+    long getKey();
 
     /**
      * 获取对应的 value 信息
